refactor: extract amount formatting to a shared util function
