+++ wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/screen/transaction_history/TransactionHistoryMapper.kt
@@ -1,12 +1,12 @@
 package net.thechance.mena.wallet.presentation.screen.transaction_history
 
-import kotlinx.datetime.LocalDate
 import net.thechance.mena.wallet.domain.entity.Transaction
 import net.thechance.mena.wallet.domain.entity.TransactionStatus
 import net.thechance.mena.wallet.domain.entity.TransactionType
 import net.thechance.mena.wallet.domain.model.TransactionFilterParams
 import net.thechance.mena.wallet.presentation.model.FilterStatus
 import net.thechance.mena.wallet.presentation.model.FilterType
+import net.thechance.mena.wallet.presentation.utils.formatAmountWithCommas
 import net.thechance.mena.wallet.presentation.utils.formatLocalDateTime
 import kotlin.uuid.ExperimentalUuidApi
 
@@ -15,31 +15,11 @@ fun Transaction.toUi(): TransactionHistoryScreenState.TransactionHistoryUiState 
     TransactionHistoryScreenState.TransactionHistoryUiState(
         id = id,
         timeAndDate = formatLocalDateTime(date = createdAt, outputFormat = "dd MMM yyyy, h:mm a"),
-        amount = amount.formatWithCommas(),
+        amount = amount.formatAmountWithCommas(),
         type = getTransactionType(),
         status = getTransactionStatus(),
         contactName = getUserName()
     )
-
-fun Double.formatWithCommas(): String {
-    val roundedAmount = (this * 100).toLong() / 100.0
-    val intPart = roundedAmount.toLong()
-    val decimalPart = ((roundedAmount - intPart) * 100).toInt()
-
-    return if (this >= 1000) {
-        val intString = intPart.toString()
-        val withCommas = buildString {
-            intString.reversed().forEachIndexed { index, char ->
-                if (index > 0 && index % 3 == 0) append(',')
-                append(char)
-            }
-        }.reversed()
-
-        "$withCommas.${decimalPart.toString().padStart(2, '0')}"
-    } else {
-        "$intPart.${decimalPart.toString().padStart(2, '0')}"
-    }
-}
 
 fun TransactionFilterState.toParams(): TransactionFilterParams {
     return TransactionFilterParams(
